feat: 더보기 버튼 클릭시 추가 적인 댓글이 보이도록 구현

+++ src/main/java/kr/codesqaud/cafe/reply/repository/ReplyRepository.java
@@ -15,5 +15,5 @@ public interface ReplyRepository {
 
 	String findReplyIdByIdx(Long replyIdx);
 
-	int getCountOfReplies(Long articleIdx);
+	Long getCountOfReplies(Long articleIdx);
 }
